% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR-plot.R
\name{OR_grid}
\alias{OR_grid}
\title{Generate a grid of Odds Ratios}
\usage{
OR_grid(dat, x, y, OR = "OR", p = "p", alpha = 0.05, breaks = c(0, 0.3,
  0.7, 1, 2, 5, Inf), label_size = NULL, coord_flip = FALSE)
}
\arguments{
\item{dat}{The input data}

\item{x}{Column mapping to x}

\item{y}{Column mapping to y}

\item{OR}{Column containing odds ratios}

\item{p}{Column in p-values}

\item{alpha}{The significance cutoff}

\item{breaks}{Break points for color-coding effect sizes}

\item{label_size}{The size of labels indicating effect size}

\item{coord_flip}{Whether or not to flip the coordinates}
}
\value{
Returns a grid color-coded and labeled by effect size
and indicating significance
}
\description{
A function to compare odds ratios, visualize their effect size, and
detect significance between different testing conditions.
}
\seealso{
\code{\link[ggsensitivity]{OR_grid_large}}
}
