% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR-plot.R
\name{OR_grid_large}
\alias{OR_grid_large}
\title{Generate an htmlTable from tidy data}
\usage{
OR_grid_large(dat, x, y, OR = "OR", p = "p", alpha = 0.05, breaks = c(0,
  0.3, 0.7, 1, 2, 5, Inf), label_size = NULL, coord_flip = FALSE)
}
\arguments{
\item{dat}{The input data}

\item{x}{Column mapping to x}

\item{y}{Column mapping to y}

\item{OR}{Column containing odds ratios}

\item{p}{Column in p-values}

\item{alpha}{The significance cutoff}

\item{breaks}{Break points for color-coding effect sizes}

\item{label_size}{The size of labels indicating effect size}

\item{coord_flip}{Whether or not to flip the coordinates}
}
\value{
Returns a grid color-coded by effect size indicating significance
with red asterix. This table is better suited to presenting larger numbers
of outcomes than \code{OR_grid}
}
\description{
A wrapper script for \code{htmlTable} that will build an htmlTable using
tidy data and mapping elements of the table to specific columns.
}
\seealso{
\code{\link[ggsensitivity]{OR_grid}}
}
